Add option to automatically start dev tools, and make HMR activate sooner.
